Add Enter-PSHostSession cmdlet with shared base class architecture
- Created PSHostSessionCommandBase abstract base class to share parameters and logic
- Refactored NewPSHostSessionCommand to extend base class (283 lines of code eliminated)
- Added EnterPSHostSessionCommand that creates and enters sessions interactively
- Supports all 6 parameter sets: Subprocess, TCP, WebSocket, NamedPipe, ProcessId, SSH
- Updated module manifests to export Enter-PSHostSession cmdlet
- Updated documentation (README.md and copilot-instructions.md)
- All 79 existing tests pass with no regressions
