+++ src/AwakeCoding.PSRemoting.psd1
@@ -74,7 +74,7 @@ RequiredAssemblies = @(
 FunctionsToExport = @()
 
 # Cmdlets to export from this module, for best performance, do not use wildcards and do not delete the entry, use an empty array if there are no cmdlets to export.
-CmdletsToExport = @('New-PSHostSession', 'Connect-PSHostProcess', 'Start-PSHostServer', 'Stop-PSHostServer', 'Get-PSHostServer')
+CmdletsToExport = @('New-PSHostSession', 'Enter-PSHostSession', 'Connect-PSHostProcess', 'Start-PSHostServer', 'Stop-PSHostServer', 'Get-PSHostServer')
 
 # Variables to export from this module
 VariablesToExport = @()
